Move encoder to utils (#100)
